+++ application-load-balancer/locals.tf
@@ -30,7 +30,9 @@ locals {
   additional_address_fqdn = try({ for k in var.config.additional_address_list : "${k}.${local.additional_address_domain}" => "${var.application}.${local.domain_suffix}" }, {})
 
   # A List of domains that can be used in the Subject Alternative Name (SAN) part of the certificate.
-  san_list = merge(local.additional_address_fqdn, var.config.cdn_domains_list)
+  # Only select the domain from the value field of cdn_domain_list (drop "internal") 
+  culled_san_list = try({ for k, v in var.config.cdn_domains_list : k => v[1] }, {})
+  san_list        = merge(local.additional_address_fqdn, local.culled_san_list)
 
   # Create a complete domain list, primary domain plus all CDN/SAN domains.
   full_list = merge({ (local.domain_name) = "${var.application}.${local.domain_suffix}" }, local.san_list)