+++ environment-pipelines/locals.tf
@@ -16,20 +16,31 @@ locals {
         stage_name : "Plan-${env.name}",
         env : env.name,
         accounts : env.accounts,
+        input_artifacts : ["build_output"],
+        output_artifacts : ["${env.name}_terraform_plan"],
         configuration : {
           ProjectName : "${var.application}-environment-pipeline-plan"
           PrimarySource : "build_output"
-          EnvironmentVariables : jsonencode([{ name : "ENVIRONMENT", value : env.name }, { name : "COPILOT_PROFILE", value : env.accounts.deploy.name }])
+          EnvironmentVariables : jsonencode([
+            { name : "ENVIRONMENT", value : env.name },
+            { name : "COPILOT_PROFILE", value : env.accounts.deploy.name },
+            { name : "SLACK_CHANNEL_ID", value : var.slack_channel, type : "PARAMETER_STORE" },
+            { name : "SLACK_REF", value : "#{slack.SLACK_REF}" },
+            { name : "NEEDS_APPROVAL", value : coalesce(env.requires_approval, false) ? "yes" : "no" }
+          ])
         }
+        namespace : "${env.name}-plan"
       },
       # The second element of the inner list for an env is the Approval stage if required, or the empty list otherwise.
       coalesce(env.requires_approval, false) ? [{
         type : "approve",
         stage_name : "Approve-${env.name}",
-        env : ""
+        env : "",
+        input_artifacts : [],
+        output_artifacts : [],
         configuration : {
           CustomData : "Review Terraform Plan"
-          ExternalEntityLink : "https://${data.aws_region.current.name}.console.aws.amazon.com/codesuite/codebuild/${data.aws_caller_identity.current.account_id}/projects/${var.application}-environment-pipeline-tf-plan/build/#{tf-plan.BUILD_ID}"
+          ExternalEntityLink : "https://${data.aws_region.current.name}.console.aws.amazon.com/codesuite/codebuild/${data.aws_caller_identity.current.account_id}/projects/${var.application}-environment-pipeline-plan/build/#{${env.name}-plan.BUILD_ID}"
         },
         namespace : null
       }] : [],
@@ -39,10 +50,16 @@ locals {
         env : env.name,
         stage_name : "Apply-${env.name}",
         accounts : env.accounts,
+        input_artifacts : ["build_output", "${env.name}_terraform_plan"],
+        output_artifacts : [],
         configuration : {
           ProjectName : "${var.application}-environment-pipeline-apply"
           PrimarySource : "build_output"
-          EnvironmentVariables : jsonencode([{ name : "ENVIRONMENT", value : env.name }])
+          EnvironmentVariables : jsonencode([
+            { name : "ENVIRONMENT", value : env.name },
+            { name : "SLACK_CHANNEL_ID", value : var.slack_channel, type : "PARAMETER_STORE" },
+            { name : "SLACK_REF", value : "#{slack.SLACK_REF}" },
+          ])
         },
         namespace : null
       }