+++ extensions/main.tf
@@ -64,6 +64,19 @@ module "alb" {
   config = each.value
 }
 
+module "cdn" {
+  source = "../cdn"
+
+  for_each = local.cdn
+  providers = {
+    aws.domain-cdn = aws.domain-cdn
+  }
+  application = var.args.application
+  environment = var.environment
+
+  config = each.value
+}
+
 module "monitoring" {
   source = "../monitoring"
 
@@ -78,6 +91,7 @@ module "monitoring" {
 
 resource "aws_ssm_parameter" "addons" {
   name  = "/copilot/applications/${var.args.application}/environments/${var.environment}/addons"
+  tier  = "Intelligent-Tiering"
   type  = "String"
   value = jsonencode(var.args.services)
   tags  = local.tags